Implement DataProvider and Dataset separation
InfluxBackedDataset and now DataLakeBackedDataset exposed a neeed to
treat the Dataset separate from where the data itself was coming from.

We now use dependency injection from Dataset to use a 'data_provider' to
fetch the data itself. And Dataset object will have their own properties
specific to Dataset operations. Data providers main entry is .load_dataframes()
which returns an interable of single column dataframes suitable for
GordoBaseDataset.join_timeseries(...)
